Encapsulate all calls to `gsutil`
