+++ dbcrossbarlib/src/clouds/gcloud/mod.rs
@@ -1,0 +1,3 @@
+//! Interfaces to Google Cloud.
+
+pub(crate) mod storage;
