+++ dbcrossbarlib/src/clouds/mod.rs
@@ -1,0 +1,3 @@
+//! Interfaces to various clouds.
+
+pub(crate) mod gcloud;
