+++ dbcrossbarlib/src/drivers/gs/write_local_data.rs
@@ -1,11 +1,8 @@
 //! Writing data to Google Cloud Storage.
 
-use std::process::Stdio;
-use tokio::process::Command;
-
 use super::{prepare_as_destination_helper, GsLocator};
+use crate::clouds::gcloud::storage;
 use crate::common::*;
-use crate::tokio_glue::copy_stream_to_writer;
 
 /// Implementation of `write_local_data`, but as a real `async` function.
 pub(crate) async fn write_local_data_helper(
@@ -31,31 +28,8 @@ pub(crate) async fn write_local_data_helper(
             let ctx = ctx
                 .child(o!("stream" => stream.name.clone(), "url" => url.to_string()));
 
-            // Run `gsutil cp - $URL` as a background process.
-            debug!(ctx.log(), "uploading stream to gsutil");
-            let mut child = Command::new("gsutil")
-                .args(&["cp", "-", url.as_str()])
-                .stdin(Stdio::piped())
-                // Throw away stdout so it doesn't corrupt our output.
-                .stdout(Stdio::null())
-                .spawn()
-                .context("error running gsutil")?;
-            let child_stdin = child.stdin.take().expect("child should have stdin");
-
-            // Copy data to our child process.
-            copy_stream_to_writer(ctx.clone(), stream.data, child_stdin)
-                .await
-                .context("error copying data to gsutil")?;
-
-            // Wait for `gsutil` to finish.
-            let status = child
-                .await
-                .with_context(|_| format!("error finishing upload to {}", url))?;
-            if status.success() {
-                Ok(GsLocator { url }.boxed())
-            } else {
-                Err(format_err!("gsutil returned error: {}", status))
-            }
+            storage::upload_file(ctx.clone(), stream.data, &url).await?;
+            Ok(GsLocator { url }.boxed())
         }
         .boxed()
     });
